talk_to_docs_full_project_AI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
langchain==0.0.350
transformers==4.35.2
torch==2.1.1
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
streamlit==1.29.0
pytest==7.4.3
httpx==0.25.2
python-ftp==0.1.0
aiofiles==23.2.1
tenacity==8.2.3
loguru==0.7.2
prometheus-client==0.19.0
redis==5.0.1
celery==5.3.6
alembic==1.12.1
flower==2.0.1  # Celery monitoring 